library/requirements.txt
ansicon==1.89.0
asgiref==3.2.10
blessed==1.17.10
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-crispy-forms==1.9.2
django-datatables-view==1.19.1
django-datatables-views==0.3
django-import-export==2.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
jinxed==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
typesentry==0.2.7
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
